package: Rename src->server and public->client.
