+++ client/compose_box.js
