+++ client/split_pane.js
