+++ client/tab_bar.js
