+++ package.json
@@ -4,7 +4,7 @@
   "description": "[WIP] A proxy server for zulip api",
   "main": "index.js",
   "scripts": {
-    "start": "node src/index.js",
+    "start": "node server/index.js",
     "lint": "prettier . --write",
     "test": "prettier . --check && node tests/*.js"
   },
