+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 
 module.exports = {
     entry: {
-        index: path.resolve(__dirname, 'public', 'index.js'),
+        index: path.resolve(__dirname, 'client', 'index.js'),
     },
     mode: 'development',
     // resolve: {
